refactor(tui): extract footer, palettes, and plan panel components
